requirements.txt
aiogram==3.21.0
python-dotenv==1.1.1
requests==2.32.4
httpx==0.28.1
configparser==7.2.0
